% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display accidents map by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An Integer with the State Code
in case of incorrect state number provided function will rise the error}

\item{year}{A string, or an integer, with the year input}
}
\value{
map plot
}
\description{
Displays accidents (using latitude and longitude coordinates provided in dataset) at states map plot
}
\examples{
\dontrun{
  fars_map_state(17, 2013)
}
}
