% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads in data files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name/path to file to read
function will rise an error if input data file doesn't exist or path to it is incorrectly specified}
}
\value{
A data frame tbl with data readed from the csv file, or an error if the
  file does not exists.
}
\description{
Reads data from .csv files from the US
National Highway Traffic Safety Administration's Fatality Analysis
Reporting System (FARS),
- public yearly data, regarding fatal injuries suffered in motor vehicle traffic crashes
and converts them into data frame tbl
}
\examples{
\dontrun{
  library(dplyr)
  library(readr)
  yr <- 2014
  data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)
}
}
