% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data in year by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of years}
}
\value{
A data frame tbl including entries in data by month, or NULL if the
year is not valid
}
\description{
Function to read data given specified set of years
}
\examples{
\dontrun{
  fars_read_years(2013)
}
}
