% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes proper filename (path to file) given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{character or integer value for a year
function will rise an error if no data found for specified year}
}
\value{
This function returns a string with the data file name for a given year
}
\description{
Makes input data file name using correct name template 'accident_<year>.csv.bz2'
}
\examples{
\dontrun{
  make_filename(2013)
}

}
